map reading done, just need more cells and a tweak during gen
