+++ The_LabyRun/Assembly-CSharp.csproj
@@ -64,9 +64,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Assets\Script\Map\Generation.cs" />
-    <Compile Include="Assets\Script\Map\map_lauch.cs" />
     <Compile Include="Assets\Script\Map\obstacle_change.cs" />
     <Compile Include="Assets\Script\Map\reading_construction.cs" />
+    <Compile Include="Assets\Script\Map\test_map_lauch.cs" />
     <Compile Include="Assets\Script\Mouvement\Monster.cs" />
     <Compile Include="Assets\Script\Mouvement\Player.cs" />
     <Reference Include="Unity.Timeline.Editor">
