+++ The_LabyRun/Assets/Script/Map/test_map_lauch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46a370839736b1e4690ad4b24a3a2a5c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
