--- The_LabyRun/Library/EditorInstance.json
@@ -1,6 +1,0 @@
-{
-	"process_id" : 24928,
-	"version" : "2019.3.9f1",
-	"app_path" : "D:/Unity/2019.3.9f1/Editor/Unity.exe",
-	"app_contents_path" : "D:/Unity/2019.3.9f1/Editor/Data"
-}
